app/requirements.txt
fastapi==0.115.2
uvicorn[standard]
torch
numpy
Pillow
boto3
requests       
python-multipart 
openai
pydantic
scikit-learn
torchvision
torchaudio
transformers
pillow
tqdm
matplotlib
japanize-matplotlib
typer
pandas
